DV-66 feat: types, data 폴더 추가 답변 형식 재정의
