+++ src/app/interview/ongoing/page.tsx
@@ -1,88 +1,32 @@
 "use client";
+import { QUESTIONS } from "@/data/questions";
 import { useState } from "react";
 
-const QUESTIONS: { questions: Question[] } = {
-  questions: [
-    {
-      question_id: 1,
-      question_excerpt: "할 수 있다는 믿음으로 달성한 첫 수상",
-      question_text:
-        "이전에 수상을 목표로 한 대회에 참가하면서 부딪혔던 어려움과 그것을 극복하기 위한 노력에 대해 설명해주세요.",
-      question_intent:
-        "지원자가 높은 목표를 달성하기 위해 어떻게 노력하고 어려움을 극복했는지에 대해 알아봅니다.",
-      key_terms: ["목표 설정", "노력", "어려움 극복", "자기 믿음"],
-    },
-    {
-      question_id: 2,
-      question_excerpt:
-        "가장 단순한 해결 방법을 통해 효율적인 결과를 도출하다.",
-      question_text:
-        "전선 탐지 알고리즘 프로젝트에서 새로운 방식을 도입하여 문제를 해결한 경험에 대해 이야기해주세요.",
-      question_intent:
-        "지원자가 어떻게 창의적인 방법으로 문제를 해결하였는지에 대해 알아봅니다.",
-      key_terms: ["창의적 해결", "문제 해결 방법", "새로운 방식", "성과"],
-    },
-    {
-      question_id: 3,
-      question_excerpt: "소통의 오류를 문화의 이해로 해결하다",
-      question_text:
-        "일본인 교환학생과의 활동을 통해 문화적 차이를 극복하고 함께 성장한 경험에 대해 이야기해주세요.",
-      question_intent:
-        "지원자가 다양한 문화 간 소통에서 어떻게 어려움을 극복하고 상호 이해를 이루었는지에 대해 알아봅니다.",
-      key_terms: ["다문화 경험", "문화 이해", "의사소통", "문화 간 이해"],
-    },
-    {
-      question_id: 4,
-      question_excerpt: "자동화 프로그램을 활용한 효율적인 데이터 수집",
-      question_text:
-        "안구 추적 데이터 수집 프로젝트에서 Digital Transformation을 통해 어떻게 자동화 시스템을 구축하였는지에 대해 상세히 설명해주세요.",
-      question_intent:
-        "지원자가 어떻게 기존의 비효율적인 방식을 개선하고 자동화를 통해 효율성을 증대시켰는지에 대해 알아봅니다.",
-      key_terms: [
-        "Digital Transformation",
-        "자동화 시스템",
-        "효율성 증대",
-        "시스템 구축",
-      ],
-    },
-  ],
-};
-
-interface Question {
-  question_id: number;
-  question_excerpt: string;
-  question_text: string;
-  question_intent: string;
-  key_terms: string[];
-}
-
-interface InterviewOngoingPageProps {
-  questions: Question[];
-}
-
-const InterviewOngoingPage = ({
-  questions = QUESTIONS.questions,
-}: InterviewOngoingPageProps) => {
+const InterviewOngoingPage = () => {
+  const questions = QUESTIONS.questions;
   const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
-  const [answers, setAnswers] = useState({});
+  const [answers, setAnswers] = useState<
+    {
+      quetion_id: number;
+      answer_text: string;
+    }[]
+  >([]);
   const [answerText, setAnswerText] = useState("");
 
   const handleNextQuestion = () => {
-    // 현재 질문의 답변을 저장
-    setAnswers((prev) => ({
+    setAnswers((prev) => [
       ...prev,
-      [questions[currentQuestionIndex].question_id]: answerText,
-    }));
+      {
+        quetion_id: questions[currentQuestionIndex].question_id,
+        answer_text: answerText,
+      },
+    ]);
 
-    // 마지막 질문이면 백엔드로 답변 전송
     if (currentQuestionIndex === questions.length - 1) {
-      // 예시: 백엔드로 답변 전송
       console.log("Submit answers to backend:", answers);
-      // 추가: 백엔드로 POST 요청하는 로직
     } else {
-      // 다음 질문으로 이동
       setCurrentQuestionIndex((prev) => prev + 1);
-      setAnswerText(""); // 답변 입력 초기화
+      setAnswerText("");
     }
   };
 
@@ -94,8 +38,9 @@ const InterviewOngoingPage = ({
       </div>
 
       <div className="flex flex-col w-[900px] gap-6">
-        <div className="flex w-full bg-primary text-white rounded-md items-center justify-center py-3 text-lg">
-          Q. {questions[currentQuestionIndex].question_text}
+        <div className="flex w-full bg-primary text-white rounded-md items-center justify-center p-3 text-lg">
+          Q{currentQuestionIndex + 1}.{" "}
+          {questions[currentQuestionIndex].question_text}
         </div>
         <textarea
           className="w-full h-72 border-2 rounded-md p-3"
