+++ src/types/types.d.ts
@@ -1,0 +1,7 @@
+interface Question {
+  question_id: number;
+  question_excerpt: string;
+  question_text: string;
+  question_intent: string;
+  key_terms: string[];
+}
